feat(waiting-text): extract WaitingText component for better reusability
